Get-DiskSpeed skips test on ReadOnly
The WinSAT.exe will pop up an error message or rather block script processing if it is applied against a non-writable partition (for example USB stick). Therefore I added some logic to the Get-DiskSpeed function. In case of a readonly disk it will return 0 for both ReadRandom and WriteRandom.

+++ AutomatedLabDefinition/AutomatedLabDefinition.psm1
@@ -2798,18 +2798,29 @@ function Get-DiskSpeed
         $labSources = Get-LabSourcesLocation
     }
 
-    Write-ScreenInfo -Message "Measuring speed of drive $DriveLetter" -Type Info
+    $IsReadOnly = Get-Partition -DriveLetter ($DriveLetter.TrimEnd(':')) | Select-Object -ExpandProperty IsReadOnly
+    if ($IsReadOnly)
+    {
+        Write-ScreenInfo -Message "Drive $DriveLetter is read-only. Skipping disk speed test" -Type Warning
+
+        $readThroughoutRandom = 0
+        $writeThroughoutRandom = 0
+    }
+    else
+    {
+        Write-ScreenInfo -Message "Measuring speed of drive $DriveLetter" -Type Info
     
-    $tempFileName = [System.IO.Path]::GetTempFileName()
+        $tempFileName = [System.IO.Path]::GetTempFileName()
     
-    & "$labSources\Tools\WinSAT.exe" disk -ran -read -count $Interations -drive $DriveLetter -xml $tempFileName | Out-Null
-    $readThroughoutRandom = (Select-Xml -Path $tempFileName -XPath '/WinSAT/Metrics/DiskMetrics/AvgThroughput').Node.'#text'
+        & "$labSources\Tools\WinSAT.exe" disk -ran -read -count $Interations -drive $DriveLetter -xml $tempFileName | Out-Null
+        $readThroughoutRandom = (Select-Xml -Path $tempFileName -XPath '/WinSAT/Metrics/DiskMetrics/AvgThroughput').Node.'#text'
     
-    & "$labSources\Tools\WinSAT.exe" disk -ran -write -count $Interations -drive $DriveLetter -xml $tempFileName | Out-Null
-    $writeThroughoutRandom = (Select-Xml -Path $tempFileName -XPath '/WinSAT/Metrics/DiskMetrics/AvgThroughput').Node.'#text'
+        & "$labSources\Tools\WinSAT.exe" disk -ran -write -count $Interations -drive $DriveLetter -xml $tempFileName | Out-Null
+        $writeThroughoutRandom = (Select-Xml -Path $tempFileName -XPath '/WinSAT/Metrics/DiskMetrics/AvgThroughput').Node.'#text'
     
-    Remove-Item -Path $tempFileName
-    
+        Remove-Item -Path $tempFileName
+    }
+
     $result = New-Object PSObject -Property ([ordered]@{
             ReadRandom = $readThroughoutRandom
             WriteRandom = $writeThroughoutRandom
